pyproject.toml
[project]
name = "open-agora"
version = "0.1.0"
description = "Open Agora - AI Agent Marketplace with verified capabilities and x402 payments"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "motor>=3.3.2",
    "pymongo>=4.6.1",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "voyageai>=0.2.3",
    "openai>=1.12.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "x402>=0.1.0",
    "eth-account>=0.11.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "thesys_genui_sdk>=0.1.2",
    "crayonai_stream>=0.6.2",
    "mcp[cli]>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
]

[project.scripts]
bazaar = "bazaar.cli:app"
bazaar-mcp = "bazaar.mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bazaar"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
